requirements.txt

fastapi
uvicorn
jinja2
aiofiles
requests
python-dotenv
